renamed test folder structure - moved mock_pika to test_utils
